chore: rename "versions" dir to "constants"
