+++ .vitepress/constants/arabica_versions.js
@@ -1,0 +1,9 @@
+const arabicaVersions = Object.freeze({
+  "app-latest-tag": "v1.1.0",
+  "app-latest-sha": "2b8cc9e23826ccb658b7dd5aa6cd51a0921a0c29",
+  "core-latest-tag": "v1.29.0-tm-v0.34.29",
+  "core-latest-sha": "e67de4517780e8db93c74d8eec4e9c66edffbc99",
+  "node-latest-tag": "v0.11.0",
+  "node-latest-sha": "0b854485d4ee7d84ffdaf8c2e8bd1fcedf940680",
+});
+export default arabicaVersions;
