+++ .vitepress/constants/constants.js
@@ -1,0 +1,18 @@
+const constants = Object.freeze({
+  golangNodeBSR: "1.21.1",
+  golangNodeMocha: "1.21.1",
+  golangNodeArabica: "1.21.1",
+  golangApp: "1.21.1",
+  golangCore: "1.21.1",
+  golang: "1.21.1",
+  arabicaChainId: "arabica-10",
+  bsrChainId: "blockspacerace-0",
+  mochaChainId: "mocha-4",
+  arabicaRollkitVersion: "v0.10.5",
+  mochaRollkitVersion: "v0.10.5",
+  bsrRollkitVersion: "v0.8.1",
+  localCelestiaDevnetVersion: "v0.8.2",
+  golangBlobstream: "1.20.2",
+  orchrelayVersion: "v0.2.0-app-v0.13.2-beta",
+});
+export default constants;
