--- .vitepress/versions/arabica_versions.js
@@ -1,9 +1,0 @@
-const arabicaVersions = Object.freeze({
-  "app-latest-tag": "v1.0.0-rc7",
-  "app-latest-sha": "2ad80bc4aa1be3ddae6b417189ce58088a93c666",
-  "core-latest-tag": "v1.23.0-tm-v0.34.28",
-  "core-latest-sha": "67cc27bcb051f911c40444530f4e5257b04dd469",
-  "node-latest-tag": "v0.11.0-rc8",
-  "node-latest-sha": "e41f62662cf1844ba52d7a93d03cf79b6e363897",
-});
-export default arabicaVersions;
