--- .vitepress/versions/blockspacerace_versions.js
@@ -1,9 +1,0 @@
-const blockspaceraceVersions = Object.freeze({
-  "app-latest-tag": "v0.13.3",
-  "app-latest-sha": "ab64b67797653e99691e846a4303bd71f44d1a8e",
-  "core-latest-tag": "v1.21.2-tm-v0.34.27",
-  "core-latest-sha": "d280f37a8376ed54ae03b10896fa25a4cbbc6d5b",
-  "node-latest-tag": "v0.10.4",
-  "node-latest-sha": "03ff94a7d779caf1225f3dccb53a68e8f1646dc6",
-});
-export default blockspaceraceVersions;
