--- .vitepress/versions/constants.js
@@ -1,18 +1,0 @@
-const constants = Object.freeze({
-  golangNodeBSR: "1.20.2",
-  golangNodeMocha: "1.20.2",
-  golangNodeArabica: "1.20.2",
-  golangApp: "1.20.2",
-  golangCore: "1.20.2",
-  golang: "1.20.2",
-  arabicaChainId: "arabica-9",
-  bsrChainId: "blockspacerace-0",
-  mochaChainId: "mocha-3",
-  arabicaRollkitVersion: "v0.10.2",
-  mochaRollkitVersion: "v0.10.2",
-  bsrRollkitVersion: "v0.8.1",
-  localCelestiaDevnetVersion: "v0.11/0-rc8",
-  golangQGB: "1.20.2",
-  orchrelayVersion: "v0.2.0-app-v0.13.2-beta",
-});
-export default constants;
