--- .vitepress/versions/mocha_versions.js
@@ -1,9 +1,0 @@
-const mochaVersions = Object.freeze({
-  "app-latest-tag": "v1.0.0-rc10",
-  "app-latest-sha": "96b86ba8c0a11b1b854da1c256c5d737dbaeede3",
-  "core-latest-tag": "v1.24.0-tm-v0.34.28",
-  "core-latest-sha": "d24c81afe0ecc4ccfbb7eb543a45a11794f04db4",
-  "node-latest-tag": "v0.11.0-rc8",
-  "node-latest-sha": "e41f62662cf1844ba52d7a93d03cf79b6e363897",
-});
-export default mochaVersions;
