+++ developers/cosmostation.md
@@ -6,7 +6,7 @@ description: How you can add Celestia network parameters to Cosmostation wallet.
 
 <!-- markdownlint-disable MD033 -->
 <script setup>
-import { versions } from '/.vitepress/versions/data.js'
+import { versions } from '/.vitepress/constants/data.js'
 </script>
 
 This guide will go over how you can add Celestia network parameters
